apps/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentforge-api"
version = "0.1.0"
description = "AgentForge API control plane"
readme = "../../README.md"
requires-python = "==3.12.*"
license = { text = "MIT" }
authors = [{ name = "Mehul Upase" }]
urls = { repository = "https://github.com/Mehulupase01/AgentForge-Multi-Tool-Agent-with-MCP--NeMo-Guardrails---Audit-Logging" }
dependencies = [
  "fastapi==0.115.5",
  "uvicorn[standard]==0.32.1",
  "sqlalchemy==2.0.36",
  "alembic==1.14.0",
  "asyncpg==0.30.0",
  "aiosqlite==0.20.0",
  "pydantic==2.11.0",
  "pydantic-settings==2.7.0",
  "python-dotenv==1.0.1",
  "structlog==24.4.0",
  "langgraph==0.2.61",
  "langgraph-checkpoint-sqlite==2.0.11",
  "langchain-core==0.3.25",
  "langchain-openai==0.2.12",
  "openai==1.58.1",
  "mcp==1.27.0",
  "nemoguardrails==0.11.0",
  "numpy==1.26.4",
  "presidio-analyzer==2.2.355",
  "presidio-anonymizer==2.2.355",
  "spacy==3.7.6",
  "httpx==0.28.1",
  "jinja2==3.1.4",
  "sse-starlette==2.2.1",
  "click==8.1.7",
  "faker==33.1.0",
  "markdown-it-py==3.0.0",
  "sqlparse==0.5.3",
  "prometheus-client==0.21.1",
  "tiktoken==0.8.0",
  "pyyaml==6.0.2",
]

[project.scripts]
agentforge-api = "agentforge.main:main"
agentforge = "agentforge.tools.cli:main"

[dependency-groups]
test = [
  "pytest==8.3.4",
  "pytest-asyncio==0.24.0",
  "pytest-cov==6.0.0",
  "respx==0.21.1",
  "anyio==4.7.0",
]

[tool.uv]
default-groups = ["test"]

[tool.hatch.build.targets.wheel]
packages = ["src/agentforge"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
